s3ext: able to write a bulk of rows from gpdb onto s3
add s3key_writer which introduces chunkbuffer to hold intermediate rows
and uploads the chunkbuffer to S3 if it is full or the request is the last call.

+++ gpAux/extensions/gps3ext/test/s3restful_service_test.cpp
@@ -158,7 +158,7 @@ TEST(S3RESTfulService, DISABLED_PutToDummyServer) {
     data.push_back(0);
 
     headers.Add(CONTENTTYPE, "text/plain");
-    headers.Add(CONTENTLENGTH, std::to_string(data.size()));
+    headers.Add(CONTENTLENGTH, std::to_string((unsigned long long)data.size()));
 
     url = "http://localhost:8553";
